requirements.txt
langchain==0.1.0
langchain-openai==0.0.2
faiss-cpu==1.7.4
python-dotenv==1.0.0